refact(plot): move PlotMixin base-->plot module to group edit
