+++ graphkit/base.py
@@ -5,52 +5,7 @@ try:
 except ImportError:
     import collections as abc
 
-
-class PlotMixin(object):
-    """
-    Classes wishing to plot their graphs should inherit this and ...
-
-    implement property ``_plotter`` to return a "partial" callable that somehow
-    ends up calling  :func:`plot.plot_graph()` with the `graph` or any other
-    args binded appropriately.
-    The purpose is to avoid copying this function & documentation here around.
-    """
-
-    def plot(
-        self,
-        filename=None,
-        show=False,
-        jupyter=None,
-        **kws,
-    ):
-        """
-        :param str filename:
-            Write diagram into a file.
-            Common extensions are ``.png .dot .jpg .jpeg .pdf .svg``
-            call :func:`plot.supported_plot_formats()` for more.
-        :param show:
-            If it evaluates to true, opens the  diagram in a  matplotlib window.
-            If it equals `-1`, it plots but does not open the Window.
-        :param jupyter:
-            If it evaluates to true, return an SVG suitable to render
-            in *jupyter notebook cells* (`ipython` must be installed).
-        :param inputs:
-            an optional name list, any nodes in there are plotted
-            as a "house"
-        :param outputs:
-            an optional name list, any nodes in there are plotted
-            as an "inverted-house"
-        :param solution:
-            an optional dict with values to annotate nodes
-            (currently content not shown, but node drawn as "filled")
-
-        :return:
-            A :mod`pydot` instance
-
-        See :func:`graphkit.plot.plot_graph()` for  example code and
-        the legend of the plots.
-        """
-        return self._plotter(filename=filename, show=show, jupyter=jupyter, **kws)
+from . import plot
 
 
 class Data(object):
@@ -196,7 +151,7 @@ class Operation(object):
              self.provides)
 
 
-class NetworkOperation(Operation, PlotMixin):
+class NetworkOperation(Operation, plot.PlotMixin):
     def __init__(self, **kwargs):
         self.net = kwargs.pop('net')
         Operation.__init__(self, **kwargs)
